fix typing "e" resetting the prompt and clean code
